llm-evals/coding-exercises/requirements.txt
absl-py==2.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.12.3
bleach==6.2.0
cbor2==5.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.3.1
comm==0.2.2
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
defusedxml==0.8.0rc2
distro==1.9.0
einops==0.8.1
entrypoints==0.4
executing==2.2.0
expecttest==0.2.1
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.59.2
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.3.1
grpcio==1.71.0
grpclib==0.4.8
h11==0.16.0
h2==4.3.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hypothesis==6.130.12
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupytext==1.17.2
kiwisolver==1.4.9
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl==0.1.2
mistune==3.1.2
modal==1.2.2
mpmath==1.3.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.4.2
ninja==1.12.1
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.2.5
openai==2.8.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.4.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.6
pycparser==2.22
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==3.2.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
referencing==0.36.2
requests==2.32.3
rfc3339_validator==0.1.4
rfc3986_validator==0.1.1
rich==14.2.0
rpds-py==0.24.0
scipy==1.15.3
Send2Trash==1.8.3
setuptools==78.1.1.post0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack_data==0.6.3
standard-imghdr==3.13.0
sympy==1.13.3
synchronicity==0.10.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.1.0
typer==0.20.0
types-certifi==2021.10.8.3
types-dataclasses==0.6.6
types-python-dateutil==2.9.0.20241206
types-toml==0.10.8.20240310
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.3.0
watchfiles==1.1.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
